requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
